requirements.txt
django==3.2.20
djangorestframework==3.13.1
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0